+++ cibo/actions/finalize.py
@@ -4,7 +4,7 @@ from sqlite3 import IntegrityError
 from typing import List
 
 from cibo.actions import Action
-from cibo.models.client import Client
+from cibo.models import Client
 
 
 class Finalize(Action):
