+++ cibo/models/__init__.py
@@ -3,3 +3,9 @@
 Also contains dataclasses that are used throughout server logic, but aren't required
 to be persistent.
 """
+
+from cibo.models.__model__ import Model
+from cibo.models.client import Client, ClientLoginState
+from cibo.models.player import Player, PlayerSchema
+
+_ = Model, Client, ClientLoginState, Player, PlayerSchema
